+++ core/src/main/java/org/apache/iceberg/hadoop/HadoopStreams.java
@@ -25,6 +25,7 @@ import java.nio.ByteBuffer;
 import java.util.Arrays;
 import org.apache.hadoop.fs.FSDataInputStream;
 import org.apache.hadoop.fs.FSDataOutputStream;
+import org.apache.hadoop.fs.FSInputStream;
 import org.apache.iceberg.io.DelegatingInputStream;
 import org.apache.iceberg.io.DelegatingOutputStream;
 import org.apache.iceberg.io.PositionOutputStream;
@@ -38,7 +39,7 @@ import org.slf4j.LoggerFactory;
  *
  * <p>This class is based on Parquet's HadoopStreams.
  */
-class HadoopStreams {
+public class HadoopStreams {
 
   private HadoopStreams() {}
 
@@ -188,6 +189,50 @@ class HadoopStreams {
             Joiner.on("\n\t").join(Arrays.copyOfRange(createStack, 1, createStack.length));
         LOG.warn("Unclosed output stream created by:\n\t{}", trace);
       }
+    }
+  }
+
+  public static class WrappedSeekableInputStream extends FSInputStream
+      implements DelegatingInputStream {
+    private final SeekableInputStream inputStream;
+
+    public WrappedSeekableInputStream(SeekableInputStream inputStream) {
+      this.inputStream = inputStream;
+    }
+
+    @Override
+    public void seek(long pos) throws IOException {
+      inputStream.seek(pos);
+    }
+
+    @Override
+    public long getPos() throws IOException {
+      return inputStream.getPos();
+    }
+
+    @Override
+    public boolean seekToNewSource(long targetPos) throws IOException {
+      throw new UnsupportedOperationException("seekToNewSource not supported");
+    }
+
+    @Override
+    public int read() throws IOException {
+      return inputStream.read();
+    }
+
+    @Override
+    public int read(byte[] b, int off, int len) throws IOException {
+      return inputStream.read(b, off, len);
+    }
+
+    @Override
+    public void close() throws IOException {
+      inputStream.close();
+    }
+
+    @Override
+    public InputStream getDelegate() {
+      return inputStream;
     }
   }
 }
